Adds Ingredient editing and improves templates
Implements a window for editing ingredients, accessible from the ingredient template.

Updates the number up/down template to use events to propagate the change, rather than keeping state in the template itself.
The showed value on the number template is now updated when the value changes.

The default text box template now triggers a simple event when the text changes.

Adds a new diagram page to visualize templates.